pyproject.toml
[tool.poetry]
name = "baum-storage"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"
loguru = "^0.7.2"
greenlet = "^3.0.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
aiormq = "^6.7.7"
python-multipart = "^0.0.6"
requests = "^2.31.0"
pika = "^1.3.2"
celery = "^5.3.6"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
black = "^23.11.0"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
